+++ geotrek/core/tests/test_path_split.py
@@ -3,12 +3,15 @@ from django.test import TestCase
 from django.contrib.gis.geos import LineString, Point
 from django.conf import settings
 
+from unittest import skipIf
+
 from geotrek.common.utils import almostequal
 
 from geotrek.core.factories import PathFactory, TopologyFactory, NetworkFactory, UsageFactory
 from geotrek.core.models import Path, Topology
 
 
+@skipIf(not settings.TREKKING_TOPOLOGY_ENABLED, 'Test with dynamic segmentation only')
 class SplitPathTest(TestCase):
     def test_split_attributes(self):
         ab = PathFactory.create(name="AB", geom=LineString((0, 0), (4, 0)))
@@ -513,6 +516,7 @@ class SplitPathTest(TestCase):
         self.assertEqual(ab_2.length, 2)
 
 
+@skipIf(not settings.TREKKING_TOPOLOGY_ENABLED, 'Test with dynamic segmentation only')
 class SplitPathLineTopologyTest(TestCase):
 
     def test_split_tee_1(self):
@@ -1002,6 +1006,7 @@ class SplitPathLineTopologyTest(TestCase):
         self.assertTrue(topology.geom.equals(topogeom))
 
 
+@skipIf(not settings.TREKKING_TOPOLOGY_ENABLED, 'Test with dynamic segmentation only')
 class SplitPathPointTopologyTest(TestCase):
 
     def test_split_tee_1(self):
@@ -1370,6 +1375,7 @@ class SplitPathPointTopologyTest(TestCase):
         self.assertEqual((0.0, 0.0), (aggr_cd.start_position, aggr_cd.end_position))
 
 
+@skipIf(not settings.TREKKING_TOPOLOGY_ENABLED, 'Test with dynamic segmentation only')
 class SplitPathGenericTopologyTest(TestCase):
 
     def test_add_simple_path(self):