+++ geotrek/core/tests/test_views.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 import json
 import re
+from unittest import skipIf
 
 import mock
 from bs4 import BeautifulSoup
@@ -30,6 +31,7 @@ from geotrek.core.factories import (PathFactory, StakeFactory, TrailFactory, Com
 from geotrek.zoning.factories import CityFactory
 
 
+@skipIf(not settings.TREKKING_TOPOLOGY_ENABLED, 'Test with dynamic segmentation only')
 class MultiplePathViewsTest(AuthentFixturesTest, TestCase):
     def setUp(self):
         self.login()
@@ -86,6 +88,7 @@ class MultiplePathViewsTest(AuthentFixturesTest, TestCase):
         self.assertEqual(Path.objects.filter(pk__in=[path_1.pk, path_2.pk]).count(), 0)
 
 
+@skipIf(not settings.TREKKING_TOPOLOGY_ENABLED, 'Test with dynamic segmentation only')
 class PathViewsTest(CommonTest):
     model = Path
     modelfactory = PathFactory
@@ -285,27 +288,186 @@ class PathViewsTest(CommonTest):
         self.assertIn('error', response.json())
         self.logout()
 
+    def test_merge_fails_other_path_intersection_less_than_snapping(self):
+        """
+        Merge should fail if other path share merge intersection
+
+                          |
+                          C
+                          |
+
+        |--------A--------|-----------B-----------|
+
+        """
+        self.login()
+        path_a = PathFactory.create(name="A", geom=LineString((0, 0), (10, 0)))
+        path_b = PathFactory.create(name="B", geom=LineString((11, 0), (20, 0)))
+        PathFactory.create(name="C", geom=LineString((10, 1), (10, 10)))
+        response = self.client.post(reverse('core:merge_path'), {'path[]': [path_a.pk, path_b.pk]})
+        json_response = response.json()
+        self.assertIn('error', json_response)
+        self.assertEqual(json_response['error'], "You can't merge 2 paths with a 3rd path in the intersection")
+        self.logout()
+
     def test_merge_fails_other_path_intersection(self):
         """
         Merge should fail if other path share merge intersection
 
-        |--------A--------|-----------B-----------|
                           |
                           C
                           |
+        |--------A--------|-----------B-----------|
+
         """
         self.login()
-        path_a = PathFactory.create(name="A", geom=LineString((0, 0), (1, 0)))
-        path_b = PathFactory.create(name="B", geom=LineString((1, 0), (2, 0)))
-        PathFactory.create(name="C", geom=LineString((1, 0), (10, 10)))
+        path_a = PathFactory.create(name="A", geom=LineString((0, 0), (10, 0)))
+        path_b = PathFactory.create(name="B", geom=LineString((10, 0), (20, 0)))
+        PathFactory.create(name="C", geom=LineString((10, 0), (10, 10)))
         response = self.client.post(reverse('core:merge_path'), {'path[]': [path_a.pk, path_b.pk]})
-        self.assertIn('error', response.json())
+        json_response = response.json()
+        self.assertIn('error', json_response)
+        self.assertEqual(json_response['error'], "You can't merge 2 paths with a 3rd path in the intersection")
+        self.logout()
+
+    def test_merge_fails_other_path_intersection_2(self):
+        """
+        Merge should fail if other path share merge intersection
+
+                          |
+                          C (reversed)
+                          |
+        |--------A--------|-----------B-----------|
+
+        """
+        self.login()
+        path_a = PathFactory.create(name="A", geom=LineString((0, 0), (10, 0)))
+        path_b = PathFactory.create(name="B", geom=LineString((10, 0), (20, 0)))
+        PathFactory.create(name="C", geom=LineString((10, 10), (10, 0)))
+        response = self.client.post(reverse('core:merge_path'), {'path[]': [path_a.pk, path_b.pk]})
+        json_response = response.json()
+        self.assertIn('error', json_response)
+        self.assertEqual(json_response['error'], "You can't merge 2 paths with a 3rd path in the intersection")
+        self.logout()
+
+    def test_merge_fails_other_path_intersection_3(self):
+        """
+        Merge should fail if other path share merge intersection
+
+        |--------C--------|
+        C                 C
+        |                 |
+        |--------A--------|-----------B-----------|
+
+        """
+        self.login()
+        path_a = PathFactory.create(name="A", geom=LineString((0, 0), (10, 0)))
+        path_b = PathFactory.create(name="B", geom=LineString((10, 0), (20, 0)))
+        PathFactory.create(name="C", geom=LineString((0, 0), (0, 10), (10, 10), (10, 0)))
+        response = self.client.post(reverse('core:merge_path'), {'path[]': [path_a.pk, path_b.pk]})
+        json_response = response.json()
+        self.assertIn('error', json_response)
+        self.assertEqual(json_response['error'], "You can't merge 2 paths with a 3rd path in the intersection")
+        self.logout()
+
+    def test_merge_not_fail_draftpath_intersection(self):
+        """
+        Merge should not fail
+                          .
+                          C (draft)
+                          .
+        |--------A--------|-----------B-----------|
+
+        """
+        self.login()
+        path_a = PathFactory.create(name="A", geom=LineString((0, 0), (10, 0)))
+        path_b = PathFactory.create(name="B", geom=LineString((10, 0), (20, 0)))
+        PathFactory.create(name="C", geom=LineString((10, 0), (10, 10)), draft=True)
+        response = self.client.post(reverse('core:merge_path'), {'path[]': [path_a.pk, path_b.pk]})
+        self.assertIn('success', response.json())
+        self.logout()
+
+    def test_merge_not_fail_start_point_end_point(self):
+        """
+        Merge should not fail
+        |
+        C
+        |
+        |--------A--------|-----------B-----------|
+
+        """
+        self.login()
+        path_a = PathFactory.create(name="A", geom=LineString((0, 0), (10, 0)))
+        path_b = PathFactory.create(name="B", geom=LineString((10, 0), (20, 0)))
+        PathFactory.create(name="C", geom=LineString((0, 0), (0, 10)))
+        response = self.client.post(reverse('core:merge_path'), {'path[]': [path_a.pk, path_b.pk]})
+        self.assertIn('success', response.json())
+        self.logout()
+
+    def test_merge_not_fail_start_point_end_point_2(self):
+        """
+        Merge should not fail
+        |
+        C (reversed)
+        |
+        |--------A--------|-----------B-----------|
+
+        """
+        self.login()
+        path_a = PathFactory.create(name="A", geom=LineString((0, 0), (10, 0)))
+        path_b = PathFactory.create(name="B", geom=LineString((10, 0), (20, 0)))
+        PathFactory.create(name="C", geom=LineString((0, 10), (0, 0)))
+        response = self.client.post(reverse('core:merge_path'), {'path[]': [path_a.pk, path_b.pk]})
+        self.assertIn('success', response.json())
+        self.logout()
+
+    def test_merge_not_fail_start_point_end_point_3(self):
+        """
+        Merge should not fail
+                                                  |
+                                                  C
+                                                  |
+        |--------A--------|-----------B-----------|
+
+        """
+        self.login()
+        path_a = PathFactory.create(name="A", geom=LineString((0, 0), (10, 0)))
+        path_b = PathFactory.create(name="B", geom=LineString((10, 0), (20, 0)))
+        PathFactory.create(name="C", geom=LineString((20, 0), (20, 10)))
+        response = self.client.post(reverse('core:merge_path'), {'path[]': [path_a.pk, path_b.pk]})
+        self.assertIn('success', response.json())
+        self.logout()
+
+    def test_merge_not_fail_start_point_end_point_4(self):
+        """
+        Merge should not fail
+                                                  |
+                                                  C (reversed)
+                                                  |
+        |--------A--------|-----------B-----------|
+
+        """
+        self.login()
+        path_a = PathFactory.create(name="A", geom=LineString((0, 0), (10, 0)))
+        path_b = PathFactory.create(name="B", geom=LineString((10, 0), (20, 0)))
+        PathFactory.create(name="C", geom=LineString((20, 10), (20, 0)))
+        response = self.client.post(reverse('core:merge_path'), {'path[]': [path_a.pk, path_b.pk]})
+        self.assertIn('success', response.json())
         self.logout()
 
     def test_merge_works(self):
         self.login()
         p1 = PathFactory.create(name="AB", geom=LineString((0, 0), (1, 0)))
         p2 = PathFactory.create(name="BC", geom=LineString((1, 0), (2, 0)))
+        response = self.client.post(reverse('core:merge_path'), {'path[]': [p1.pk, p2.pk]})
+        self.assertIn('success', response.json())
+        self.logout()
+
+    def test_merge_works_other_line(self):
+        self.login()
+        p1 = PathFactory.create(name="AB", geom=LineString((0, 0), (1, 0)))
+        p2 = PathFactory.create(name="BC", geom=LineString((1, 0), (2, 0)))
+
+        PathFactory.create(name="CD", geom=LineString((2, 1), (3, 1)))
         response = self.client.post(reverse('core:merge_path'), {'path[]': [p1.pk, p2.pk]})
         self.assertIn('success', response.json())
         self.logout()
@@ -354,6 +516,7 @@ class PathViewsTest(CommonTest):
         self.logout()
 
 
+@skipIf(not settings.TREKKING_TOPOLOGY_ENABLED, 'Test with dynamic segmentation only')
 class PathKmlGPXTest(TestCase):
     def setUp(self):
         super(PathKmlGPXTest, self).setUp()
@@ -380,6 +543,7 @@ class PathKmlGPXTest(TestCase):
         self.assertEqual(self.kml_response['Content-Type'], 'application/vnd.google-earth.kml+xml')
 
 
+@skipIf(not settings.TREKKING_TOPOLOGY_ENABLED, 'Test with dynamic segmentation only')
 class DenormalizedTrailTest(AuthentFixturesTest):
     def setUp(self):
         self.trail1 = TrailFactory(no_path=True)
@@ -405,20 +569,25 @@ class DenormalizedTrailTest(AuthentFixturesTest):
             self.client.get(reverse('core:path_json_list'))
 
 
+@skipIf(not settings.TREKKING_TOPOLOGY_ENABLED, 'Test with dynamic segmentation only')
 class TrailViewsTest(CommonTest):
     model = Trail
     modelfactory = TrailFactory
     userfactory = PathManagerFactory
 
     def get_good_data(self):
-        path = PathFactory.create()
-        return {
+        good_data = {
             'name': 't',
             'departure': 'Below',
             'arrival': 'Above',
-            'comments': 'No comment',
-            'topology': '{"paths": [%s]}' % path.pk,
+            'comments': 'No comment'
         }
+        if settings.TREKKING_TOPOLOGY_ENABLED:
+            path = PathFactory.create()
+            good_data['topology'] = '{"paths": [%s]}' % path.pk
+        else:
+            good_data['geom'] = 'SRID=4326;LINESTRING (0.0 0.0, 1.0 1.0)'
+        return good_data
 
     def test_detail_page(self):
         self.login()
@@ -462,6 +631,7 @@ class TrailViewsTest(CommonTest):
         self.assertIn(trail, new_trail.trails.all())
 
 
+@skipIf(not settings.TREKKING_TOPOLOGY_ENABLED, 'Test with dynamic segmentation only')
 class TrailKmlGPXTest(TestCase):
     def setUp(self):
         super(TrailKmlGPXTest, self).setUp()
@@ -488,6 +658,7 @@ class TrailKmlGPXTest(TestCase):
         self.assertEqual(self.kml_response['Content-Type'], 'application/vnd.google-earth.kml+xml')
 
 
+@skipIf(not settings.TREKKING_TOPOLOGY_ENABLED, 'Test with dynamic segmentation only')
 class RemovePathKeepTopology(TestCase):
     def test_remove_poi(self):
         """