+++ geotrek/tourism/tests/test_models.py
@@ -23,11 +23,15 @@ class TourismRelations(TestCase):
         self.content2 = tourism_factories.TouristicContentFactory(geom='SRID=%s;POINT(2 2)' % settings.SRID)
         self.event = tourism_factories.TouristicEventFactory(geom='SRID=%s;POINT(50 50)' % settings.SRID)
         self.event2 = tourism_factories.TouristicEventFactory(geom='SRID=%s;POINT(60 60)' % settings.SRID)
-        path = core_factories.PathFactory(geom='SRID=%s;LINESTRING(0 100, 100 100)' % settings.SRID)
-        self.trek = trekking_factories.TrekFactory(no_path=True)
-        self.trek.add_path(path)
-        self.poi = trekking_factories.POIFactory(no_path=True)
-        self.poi.add_path(path, start=0.5, end=0.5)
+        if settings.TREKKING_TOPOLOGY_ENABLED:
+            path = core_factories.PathFactory(geom='SRID=%s;LINESTRING(0 100, 100 100)' % settings.SRID)
+            self.trek = trekking_factories.TrekFactory(no_path=True)
+            self.trek.add_path(path)
+            self.poi = trekking_factories.POIFactory(no_path=True)
+            self.poi.add_path(path, start=0.5, end=0.5)
+        else:
+            self.trek = trekking_factories.TrekFactory(geom='SRID=%s;LINESTRING(0 100, 100 100)' % settings.SRID)
+            self.poi = trekking_factories.POIFactory(geom='SRID=%s;POINT(50 100)' % settings.SRID)
 
     def test_spatial_link_with_tourism(self):
         self.assertIn(self.content2, self.content.touristic_contents.all())